Add the elements_include_directories macro
Dependending on the HIDE_SYSINC_WARNINGS cmake option (OFF by
default), the include directories which are not starting with one of
the CMAKE_PROJECT_PATH entries are added with the SYSTEM option. In
that case, the "-isystem" gcc command line is used instead of the
traditional "-I". The warnings coming from these directories are then
ignored.